server/core/encryption.py
"""
Encryption Handler - ChaCha20-Poly1305 AEAD Cipher
Provides encryption/decryption with authentication and replay protection.
"""

import os
import struct
import time
from typing import Optional, Tuple
from Crypto.Cipher import ChaCha20_Poly1305
from Crypto.Protocol.KDF import HKDF
from Crypto.Hash import SHA256

import sys
sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(__file__))))

from shared.protocol import KEY_SIZE, NONCE_SIZE, TAG_SIZE, KDF_INFO, KDF_SALT_SIZE
from shared.exceptions import EncryptionError, DecryptionError, ReplayAttackError


class EncryptionHandler:
    """
    Handles ChaCha20-Poly1305 AEAD encryption/decryption.
    
    Features:
    - ChaCha20-Poly1305 authenticated encryption
    - Per-packet unique nonce (counter + timestamp)
    - Replay attack protection (sliding window)
    - Key derivation from master key
    
    Thread-safe: No (create one instance per connection)
    """
    
    def __init__(self, master_key: Optional[bytes] = None):
        """
        Initialize encryption handler.
        
        Args:
            master_key: 32-byte master key. If None, generates random key.
        """
        if master_key is None:
            self.master_key = os.urandom(KEY_SIZE)
        else:
            if len(master_key) != KEY_SIZE:
                raise ValueError(f"Master key must be {KEY_SIZE} bytes")
            self.master_key = master_key
        
        # Derive session key from master key
        self.session_key = self._derive_key(self.master_key, b"session_key")
        
        # Nonce counter (8 bytes, incremented per packet)
        self.send_counter: int = 0
        self.recv_counter: int = 0
        
        # Replay protection: track received counters
        self.replay_window_size = 64
        self.highest_recv_counter = 0
        self.recv_bitmap = 0  # Bitmap of received packets in window
        
        # Statistics
        self.packets_encrypted = 0
        self.packets_decrypted = 0
        self.replay_attacks_blocked = 0
    
    def _derive_key(self, master_key: bytes, context: bytes) -> bytes:
        """
        Derive a key using HKDF-SHA256.
        
        Args:
            master_key: Master key
            context: Context for key derivation
        
        Returns:
            Derived 32-byte key
        """
        salt = KDF_INFO  # Use protocol info as salt
        derived_key = HKDF(
            master=master_key,
            key_len=KEY_SIZE,
            salt=salt,
            hashmod=SHA256,
            context=context
        )
        return derived_key
    
    def encrypt(self, plaintext: bytes) -> bytes:
        """
        Encrypt data with ChaCha20-Poly1305.
        
        Args:
            plaintext: Data to encrypt
        
        Returns:
            Encrypted packet: [8 bytes: nonce_counter] [N bytes: ciphertext] [16 bytes: tag]
        
        Raises:
            EncryptionError: If encryption fails
        """
        try:
            # Generate nonce: 8 bytes counter + 4 bytes timestamp
            nonce_counter = self.send_counter
            timestamp = int(time.time()) & 0xFFFFFFFF  # 4 bytes
            
            nonce = struct.pack('!Q', nonce_counter) + struct.pack('!I', timestamp)
            assert len(nonce) == NONCE_SIZE
            
            # Encrypt with ChaCha20-Poly1305
            cipher = ChaCha20_Poly1305.new(key=self.session_key, nonce=nonce)
            ciphertext, tag = cipher.encrypt_and_digest(plaintext)
            
            # Increment counter
            self.send_counter += 1
            self.packets_encrypted += 1
            
            # Return: [nonce_counter (8)] [ciphertext (N)] [tag (16)]
            return struct.pack('!Q', nonce_counter) + ciphertext + tag
        
        except Exception as e:
            raise EncryptionError(f"Encryption failed: {e}")
    
    def decrypt(self, encrypted_packet: bytes) -> bytes:
        """
        Decrypt ChaCha20-Poly1305 encrypted packet.
        
        Args:
            encrypted_packet: [8 bytes: nonce_counter] [N bytes: ciphertext] [16 bytes: tag]
        
        Returns:
            Decrypted plaintext
        
        Raises:
            DecryptionError: If decryption or authentication fails
            ReplayAttackError: If packet is replayed
        """
        try:
            # Minimum size: 8 (counter) + 16 (tag) = 24 bytes
            if len(encrypted_packet) < 24:
                raise DecryptionError("Packet too short")
            
            # Extract components
            nonce_counter = struct.unpack('!Q', encrypted_packet[0:8])[0]
            ciphertext = encrypted_packet[8:-16]
            tag = encrypted_packet[-16:]
            
            # Replay protection check
            if not self._check_replay(nonce_counter):
                self.replay_attacks_blocked += 1
                raise ReplayAttackError(f"Replay detected: counter={nonce_counter}")
            
            # Reconstruct nonce (we only have counter, timestamp is unknown)
            # This is OK because counter alone provides uniqueness
            # We'll use a placeholder for timestamp
            timestamp = 0  # Placeholder (not used for decryption verification)
            nonce = struct.pack('!Q', nonce_counter) + struct.pack('!I', timestamp)
            
            # Decrypt with ChaCha20-Poly1305
            cipher = ChaCha20_Poly1305.new(key=self.session_key, nonce=nonce)
            plaintext = cipher.decrypt_and_verify(ciphertext, tag)
            
            # Update replay window
            self._update_replay_window(nonce_counter)
            self.packets_decrypted += 1
            
            return plaintext
        
        except ReplayAttackError:
            raise
        except Exception as e:
            raise DecryptionError(f"Decryption failed: {e}")
    
    def _check_replay(self, counter: int) -> bool:
        """
        Check if packet counter is valid (not replayed).
        
        Args:
            counter: Packet counter
        
        Returns:
            True if packet is valid, False if replayed
        """
        # Packet is too old (outside window)
        if counter + self.replay_window_size < self.highest_recv_counter:
            return False
        
        # Packet is new (future)
        if counter > self.highest_recv_counter:
            return True
        
        # Packet is within window - check bitmap
        diff = self.highest_recv_counter - counter
        if diff >= self.replay_window_size:
            return False
        
        # Check if bit is set (already received)
        if self.recv_bitmap & (1 << diff):
            return False  # Already received
        
        return True
    
    def _update_replay_window(self, counter: int) -> None:
        """
        Update replay protection window after accepting packet.
        
        Args:
            counter: Packet counter
        """
        # New packet (future)
        if counter > self.highest_recv_counter:
            # Shift window
            diff = counter - self.highest_recv_counter
            if diff < self.replay_window_size:
                self.recv_bitmap = (self.recv_bitmap << diff) | 1
            else:
                self.recv_bitmap = 1
            self.highest_recv_counter = counter
        else:
            # Packet within window
            diff = self.highest_recv_counter - counter
            self.recv_bitmap |= (1 << diff)
    
    def rekey(self, new_master_key: bytes) -> None:
        """
        Rotate encryption keys.
        
        Args:
            new_master_key: New 32-byte master key
        
        Raises:
            ValueError: If key size is invalid
        """
        if len(new_master_key) != KEY_SIZE:
            raise ValueError(f"Master key must be {KEY_SIZE} bytes")
        
        self.master_key = new_master_key
        self.session_key = self._derive_key(self.master_key, b"session_key")
        
        # Reset counters (important!)
        self.send_counter = 0
        # Don't reset recv_counter - allow old packets to drain
    
    def get_statistics(self) -> dict:
        """
        Get encryption statistics.
        
        Returns:
            Dictionary with stats
        """
        return {
            'packets_encrypted': self.packets_encrypted,
            'packets_decrypted': self.packets_decrypted,
            'replay_attacks_blocked': self.replay_attacks_blocked,
            'send_counter': self.send_counter,
            'recv_counter': self.recv_counter,
            'highest_recv_counter': self.highest_recv_counter,
        }
    
    @staticmethod
    def generate_master_key() -> bytes:
        """Generate a random 32-byte master key."""
        return os.urandom(KEY_SIZE)
    
    @staticmethod
    def derive_key_from_password(password: str, salt: bytes) -> bytes:
        """
        Derive a master key from password.
        
        Args:
            password: User password
            salt: Random salt (should be at least 16 bytes)
        
        Returns:
            32-byte derived key
        """
        return HKDF(
            master=password.encode('utf-8'),
            key_len=KEY_SIZE,
            salt=salt,
            hashmod=SHA256,
            context=KDF_INFO
        )


# ============================================================
# Helper Functions
# ============================================================

def encrypt_packet(plaintext: bytes, key: bytes) -> bytes:
    """
    Convenience function to encrypt a single packet.
    
    Args:
        plaintext: Data to encrypt
        key: 32-byte encryption key
    
    Returns:
        Encrypted packet
    """
    handler = EncryptionHandler(master_key=key)
    return handler.encrypt(plaintext)


def decrypt_packet(encrypted: bytes, key: bytes) -> bytes:
    """
    Convenience function to decrypt a single packet.
    
    Args:
        encrypted: Encrypted packet
        key: 32-byte encryption key
    
    Returns:
        Decrypted plaintext
    """
    handler = EncryptionHandler(master_key=key)
    return handler.decrypt(encrypted)
